+++ src/applications/vaos/tests/services/appointment/index.v2.unit.spec.js
@@ -1,310 +1,21 @@
 import { expect } from 'chai';
-import omit from 'platform/utilities/data/omit';
 import sinon from 'sinon';
 import { diff } from 'just-diff';
-import { mockFetch, setFetchJSONResponse } from 'platform/testing/unit/helpers';
+import {
+  mockFetch,
+  setFetchJSONFailure,
+  setFetchJSONResponse,
+} from 'platform/testing/unit/helpers';
 
 import { fetchBookedAppointment } from '../../../services/appointment';
 import { VIDEO_TYPES } from '../../../utils/constants';
 import moment from '../../../lib/moment-tz';
-
-function getAppointmentMock({
-  id = null,
-  email = null,
-  currentStatus = null,
-  version = 0,
-  clinicFriendlyName = null,
-  clinicName = null,
-  instructionsTitle = null,
-  communityCareProvider = null,
-  timezone = null,
-  ...fields
-} = {}) {
-  const fieldsWithoutProps = omit(['email'], fields);
-  if (version === 0 && fields.kind !== 'cc') {
-    const vdsAppointments = [];
-    const vvsAppointments = [];
-    let vistaStatus = currentStatus;
-
-    if (
-      !vistaStatus &&
-      (fields.kind !== 'telehealth' || fields.vvsKind === VIDEO_TYPES.clinic)
-    ) {
-      if (fields.status === 'cancelled') {
-        vistaStatus = 'CANCELLED BY PATIENT';
-      } else if (
-        fields.status === 'booked' &&
-        moment().isSameOrAfter(fields.start)
-      ) {
-        vistaStatus = 'CHECKED OUT';
-      } else if (fields.status === 'booked') {
-        vistaStatus = 'FUTURE';
-      }
-    }
-
-    if (
-      fields.kind === 'clinic' ||
-      fields.kind === 'phone' ||
-      (fields.kind === 'telehealth' &&
-        fields.telehealth.vvsKind === VIDEO_TYPES.clinic)
-    ) {
-      vdsAppointments.push({
-        bookingNote: fields.comment,
-        appointmentLength: fields.minutesDuration,
-        appointmentTime: fields.start,
-        clinic: {
-          name: clinicName,
-          askForCheckIn: false,
-          facilityCode: fields.locationId?.substr(0, 3) || null,
-        },
-        type: 'REGULAR',
-        currentStatus: vistaStatus,
-      });
-    }
-
-    if (fields.kind === 'telehealth') {
-      vvsAppointments.push({
-        id,
-        appointmentKind: fields.telehealth.vvsKind,
-        sourceSystem: 'VCM',
-        dateTime: fields.start,
-        duration: fields.minutesDuration || null,
-        status: { description: null, code: 'FUTURE' },
-        schedulingRequestType: null,
-        type: null,
-        bookingNotes: fields.comment,
-        instructionsOther: null,
-        instructionsTitle,
-        patients: [
-          {
-            name: { firstName: 'JUDY', lastName: 'MORRISON' },
-            contactInformation: {
-              mobile: null,
-              preferredEmail: email || null,
-            },
-            location: {
-              type: 'NonVA',
-              facility: {
-                name: null,
-                siteCode: fields.locationId?.substr(0, 3),
-                timeZone: null,
-              },
-            },
-            patientAppointment: true,
-            virtualMeetingRoom: {
-              conference: null,
-              pin: null,
-              url: fields.telehealth.url,
-            },
-          },
-        ],
-        tasInfo: fields.telehealth.atlas || null,
-        providers:
-          fields.practitioners?.map(p => ({
-            name: {
-              firstName: p.firstName,
-              lastName: p.lastName,
-            },
-          })) || [],
-      });
-    }
-
-    return {
-      id,
-      type: 'appointment',
-      attributes: {
-        startDate: fields.start || null,
-        uniqueId: id,
-        clinicFriendlyName,
-        clinicId: fields.clinic || null,
-        facilityId: fields.locationId?.substr(0, 3) || null,
-        sta6aid: fields.locationId || null,
-        communityCare: false,
-        phoneOnly: fields.kind === 'phone',
-        char4: fields.serviceType === 'covid' ? 'CDQC' : null,
-        vdsAppointments,
-        vvsAppointments,
-      },
-    };
-  }
-
-  if (version === 0 && fields.kind === 'cc') {
-    return {
-      id,
-      type: 'cc_appointments',
-      attributes: {
-        appointmentRequestId: id,
-        distanceEligibleConfirmed: true,
-        name: { firstName: null, lastName: null },
-        providerPractice: communityCareProvider.name,
-        providerPhone: communityCareProvider.caresitePhone,
-        address: {
-          street: communityCareProvider.street,
-          city: communityCareProvider.city,
-          state: communityCareProvider.state,
-          zipCode: communityCareProvider.zip,
-        },
-        instructionsToVeteran: fields.comment,
-        appointmentTime: moment(fields.start)
-          .utc()
-          .format('MM/DD/YYYY HH:mm:ss'),
-        timeZone: moment(fields.start)
-          .tz(moment.tz.guess())
-          .format('Z z'),
-      },
-    };
-  }
-
-  if (version === 2) {
-    return {
-      id,
-      type: 'appointments',
-      attributes: {
-        id,
-        cancellationReason: null,
-        clinic: null,
-        comment: null,
-        contact: {
-          telecom: [
-            {
-              type: 'email',
-              value: email,
-            },
-          ],
-        },
-        description: null,
-        end: null,
-        kind: null,
-        locationId: null,
-        minutesDuration: null,
-        practitioners: communityCareProvider
-          ? [
-              {
-                id: { system: 'HSRM', value: communityCareProvider.uniqueId },
-              },
-            ]
-          : null,
-        preferredTimesForPhoneCall: null,
-        priority: null,
-        reason: null,
-        requestedPeriods: null,
-        serviceType: null,
-        slot: null,
-        start: null,
-        status: null,
-        telehealth: null,
-        ...fieldsWithoutProps,
-      },
-    };
-  }
-
-  return null;
-}
-
-function getVideoAppointmentMock({
-  id = null,
-  locationId = null,
-  vvsKind = null,
-  providers = [],
-  instructionsTitle = null,
-  start = null,
-  clinicId = null,
-  clinicName = null,
-  clinicFriendlyName = null,
-  atlasConfirmationCode = null,
-  atlasAddress = null,
-  appointmentLength = null,
-  currentStatus = null,
-  email = null,
-  url = null,
-} = {}) {
-  const vdsAppointments = [];
-
-  if (vvsKind === VIDEO_TYPES.clinic) {
-    vdsAppointments.push({
-      bookingNote: null,
-      appointmentLength,
-      appointmentTime: start,
-      clinic: {
-        name: clinicName,
-        askForCheckIn: false,
-        facilityCode: locationId?.substr(0, 3) || null,
-      },
-      type: 'REGULAR',
-      currentStatus,
-    });
-  }
-
-  const vvsAppointment = {
-    id,
-    appointmentKind: vvsKind,
-    sourceSystem: null,
-    dateTime: start,
-    duration: appointmentLength || null,
-    status: { description: null, code: 'FUTURE' },
-    schedulingRequestType: null,
-    type: null,
-    bookingNotes: null,
-    instructionsOther: null,
-    instructionsTitle,
-    patients: [
-      {
-        name: { firstName: 'JUDY', lastName: 'MORRISON' },
-        contactInformation: {
-          mobile: null,
-          preferredEmail: email || null,
-        },
-        location: {
-          type: 'NonVA',
-          facility: {
-            name: null,
-            siteCode: locationId?.substr(0, 3),
-            timeZone: null,
-          },
-        },
-        patientAppointment: true,
-        virtualMeetingRoom: {
-          conference: null,
-          pin: null,
-          url,
-        },
-      },
-    ],
-    providers,
-  };
-
-  if (atlasAddress) {
-    vvsAppointment.tasInfo = {
-      siteCode: null,
-      slotId: null,
-      confirmationCode: atlasConfirmationCode,
-      address: atlasAddress,
-      contacts: [],
-    };
-  }
-
-  return {
-    id,
-    type: 'appointment',
-    attributes: {
-      startDate: start || null,
-      clinicFriendlyName,
-      clinicId: clinicId || null,
-      facilityId: locationId?.substr(0, 3) || null,
-      sta6aid: locationId || null,
-      communityCare: false,
-      phoneOnly: false,
-      char4: null,
-      vdsAppointments,
-      vvsAppointments: [vvsAppointment],
-    },
-  };
-}
+import { createMockAppointment } from '../../mocks/data';
 
 describe('VAOS Appointment service', () => {
   describe('fetchBookedAppointment', () => {
     beforeEach(() => mockFetch());
-    it('should return an in person VA appointment', async () => {
+    it('should return matching v0 and v2 data for an in person VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -321,13 +32,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -360,7 +71,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a cancelled VA appointment', async () => {
+    it('should return matching v0 and v2 data for a cancelled VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -375,13 +86,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -412,7 +123,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a past VA appointment', async () => {
+    it('should return matching v0 and v2 data for a past VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -427,13 +138,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -460,7 +171,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a VA phone appointment', async () => {
+    it('should return matching v0 and v2 data for a VA phone appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -475,13 +186,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -508,7 +219,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a VA covid vaccine appointment', async () => {
+    it('should return matching v0 and v2 data for a VA covid vaccine appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -524,13 +235,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -557,7 +268,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a VVC at home video appointment', async () => {
+    it('should return matching v0 and v2 data for a VVC at home video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -582,13 +293,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -614,7 +325,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return an ATLAS video appointment', async () => {
+    it('should return matching v0 and v2 data for an ATLAS video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -652,13 +363,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -684,7 +395,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a clinic based video appointment', async () => {
+    it('should return matching v0 and v2 data for a clinic based video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -703,13 +414,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -735,7 +446,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a store forward video appointment', async () => {
+    it('should return matching v0 and v2 data for a store forward video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -754,13 +465,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -786,7 +497,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a mobile any video appointment', async () => {
+    it('should return matching v0 and v2 data for a mobile any video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -803,13 +514,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -836,7 +547,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a mobile gfe video appointment', async () => {
+    it('should return matching v0 and v2 data for a mobile gfe video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -853,13 +564,13 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(
           sinon.match(`/vaos/v0/appointments/va/${data.id}`),
         ),
-        { data: getAppointmentMock({ version: 0, ...data }) },
+        { data: createMockAppointment({ version: 0, ...data }) },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -886,7 +597,7 @@ describe('VAOS Appointment service', () => {
       );
     });
 
-    it('should return a community care appointment', async () => {
+    it('should return matching v0 and v2 data for a community care appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -910,11 +621,11 @@ describe('VAOS Appointment service', () => {
 
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
-        { data: getAppointmentMock({ version: 2, ...data }) },
+        { data: createMockAppointment({ version: 2, ...data }) },
       );
       setFetchJSONResponse(
         global.fetch.withArgs(sinon.match(`/vaos/v0/appointments`)),
-        { data: [getAppointmentMock({ version: 0, ...data })] },
+        { data: [createMockAppointment({ version: 0, ...data })] },
       );
 
       const [v2Result, v0Result] = await Promise.all([
@@ -954,8 +665,50 @@ describe('VAOS Appointment service', () => {
         'Transformers for v0 and v2 appointment data are out of sync',
       );
     });
-    //should return an error
+
+    it('should return matching v0 and v2 data for an error', async () => {
+      const error = {
+        code: 'VAOS_504',
+        title: 'Gateway error',
+        status: 504,
+        source: 'stack trace',
+      };
+      setFetchJSONFailure(
+        global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/1234`)),
+        {
+          errors: [error],
+        },
+      );
+      setFetchJSONFailure(
+        global.fetch.withArgs(sinon.match(`/vaos/v0/appointments/va/1234`)),
+        {
+          errors: [error],
+        },
+      );
+
+      let v0Result = null;
+      let v2Result = null;
+      try {
+        await fetchBookedAppointment({
+          id: '1234',
+          type: 'va',
+          useV2: true,
+        });
+      } catch (e) {
+        v0Result = e;
+      }
+
+      try {
+        await fetchBookedAppointment({ id: '1234', type: 'va' });
+      } catch (e) {
+        v2Result = e;
+      }
+
+      expect(v2Result).to.be.ok;
+      expect(v0Result).to.be.ok;
+
+      const differences = diff(v2Result, v0Result);
+      expect(differences).to.be.empty;
+    });
   });
-  //should return a VA request
-  //should return a community care request
 });
